style(model): Application.java에 있던 모델을 모델 패키지 만들어서 분리
변경 전
- Application 내부에 모든 모델 존재
- 각 모델 이름 끝에 Model이 들어가는 형태

변경 후
- calculator 패키지에 model 패키지를 만들고 그 안에 Calculator, Delimiter, Expression, PositiveNumber 클래스 추가 하여 기존 Application.java에서 모델 분리
- 모델 클래스 이름의 끝에 Model 삭제
