+++ src/main/java/calculator/Application.java
@@ -1,105 +1,8 @@
 package calculator;
 
+import calculator.model.Calculator;
+import calculator.model.Expression;
 import camp.nextstep.edu.missionutils.Console;
-import java.util.regex.Pattern;
-
-// model
-class PositiveNumberModel {
-    private final String REGEXP_ONLY_NUM = "^[\\d]*$";
-    private final String EMPTY_NUMBER = "";
-
-    private int count;
-
-    public PositiveNumberModel(int count) {
-        this.count = count;
-    }
-
-    public void plus(String numberAdded) {
-        if (!Pattern.matches(REGEXP_ONLY_NUM, numberAdded)) {
-            throw new IllegalArgumentException("잘못된 구분자를 포함합니다.");
-        }
-        if (numberAdded.equals(EMPTY_NUMBER)) {
-            return;
-        }
-
-        int number = Integer.parseInt(numberAdded);
-        if (number <= 0) {
-            throw new IllegalArgumentException("양수가 아닌 수를 포함합니다.");
-        }
-        this.count += number;
-    }
-
-    public int getCount() {
-        return this.count;
-    }
-}
-
-class ExpressionModel {
-    private final String expression;
-
-    public ExpressionModel(String expression) {
-        this.expression = expression;
-    }
-
-    public boolean hasCustomDelimiter() {
-        return expression.startsWith("//");
-    }
-
-    public String delimiterSection() {
-        int startIdx = expression.indexOf("//") + 2;
-        int endIdx = expression.indexOf("\\n");
-        return expression.substring(startIdx, endIdx);
-    }
-
-    public String nonDelimiterSection() {
-        int startIdx = expression.indexOf("\\n") + 2;
-        return expression.substring(startIdx);
-    }
-
-    public String getExpression() {
-        return expression;
-    }
-}
-
-class DelimiterModel {
-    private String delimiter;
-
-    public DelimiterModel() {
-        this.delimiter = ",|:";
-    }
-
-    public void registerDelimiter(String delimiterSection) {
-        for (int nowDelimiterIdx = 0; nowDelimiterIdx < delimiterSection.length(); nowDelimiterIdx++) {
-            this.delimiter += ("|\\" + delimiterSection.charAt(nowDelimiterIdx));
-        }
-    }
-
-    public String getDelimiter() {
-        return delimiter;
-    }
-}
-
-class CalculatorModel {
-    private final DelimiterModel delimiter = new DelimiterModel();
-
-    public int calculate(ExpressionModel expressionModel) {
-        String nonDelimiterSection = expressionModel.getExpression();
-
-        if (expressionModel.hasCustomDelimiter()) {
-            delimiter.registerDelimiter(expressionModel.delimiterSection());
-            nonDelimiterSection = expressionModel.nonDelimiterSection();
-        }
-
-        String[] numbers = nonDelimiterSection.split(delimiter.getDelimiter());
-
-        PositiveNumberModel sum = new PositiveNumberModel(0);
-        for (String nowNum : numbers) {
-            sum.plus(nowNum);
-        }
-        return sum.getCount();
-    }
-
-}
 
 // view
 class CalculateView {
@@ -115,10 +18,10 @@ class CalculateView {
 
 // controller
 class CalculatorController {
-    private final CalculatorModel model;
+    private final Calculator model;
     private final CalculateView view;
 
-    public CalculatorController(CalculatorModel model, CalculateView view) {
+    public CalculatorController(Calculator model, CalculateView view) {
         this.model = model;
         this.view = view;
     }
@@ -127,7 +30,7 @@ class CalculatorController {
         String expression = view.getUserInput();
 
         // 데이터 처리
-        ExpressionModel expressionModel = new ExpressionModel(expression);
+        Expression expressionModel = new Expression(expression);
         int result = model.calculate(expressionModel);
 
         // 출력
@@ -137,7 +40,7 @@ class CalculatorController {
 
 public class Application {
     public static void main(String[] args) {
-        CalculatorModel model = new CalculatorModel();
+        Calculator model = new Calculator();
         CalculateView view = new CalculateView();
         CalculatorController controller = new CalculatorController(model, view);
